requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.40.3
python-dotenv==1.0.1
lxml==5.3.0 